+++ .coveragerc
@@ -5,8 +5,9 @@ branch = True
 cover_pylib = False
 # omit uninteresting files
 omit =
-    polls/test.py
+    polls/tests/*
     polls/migrations/*
     */site-packages/*
     */dist-packages/*
     manage.py
+
