+++ requirements.txt
@@ -2,3 +2,4 @@ flake8==3.9.2
 coverage==5.5
 django==3.2.6
 python-dateutil
+python-decouple
